backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Eric <erik.lemoine@gmail.com>"]
readme = "README.md"
packages = [{include = "app", from = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
stravalib = "^2.0"
fastapi = {extras = ["standard"], version = "^0.115.2"}
sqlalchemy = "^2.0.36"
mysqlclient = "^2.2.5"
pydantic-settings = "^2.6.0"
alembic = "^1.13.3"
passlib = "^1.7.4"
argon2-cffi = "^23.1.0"
pyjwt = "^2.9.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.9"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.24.0"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
    "--cov-report=term-missing",
    "--cov-branch",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
